# docs/backend/backend-technical-spec/openapi/openapi-skeleton.yaml
openapi: 3.1.0
info:
  title: VTS EDU Backend API
  version: 0.2.0
  description: Product-ready core API for LMS and online assessments.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /me/notifications/unread-count:
    get:
      operationId: notifications.unreadCount
      tags: [Notifications]
      description: Returns the caller's current unread notification count.
      responses:
        '200':
          description: Unread count
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotificationUnreadCount'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /me/notifications/{id}/read:
    post:
      operationId: notifications.markRead
      tags: [Notifications]
      description: |
        Idempotently marks a single notification as read for the
        current caller. Notifications already in the read state are
        left unchanged.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Updated notification
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /assessments/{id}/validate:
    post:
      operationId: assessments.validate
      tags: [Assessments]
      description: Teacher or admin only. Validate that the assessment is ready to publish.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Validation result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResult'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ValidationResult:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [valid]
          properties:
            valid: { type: boolean }
            errors:
              type: array
              items:
                type: object
                required: [field, message]
                properties:
                  field: { type: string }
                  message: { type: string }
...
    Problem:
      type: object
      required: [title, status, code]
      properties:
        type: { type: string, format: uri-reference }
        title: { type: string }
        status: { type: integer }
        code: { type: string }
        detail: { type: string }
        request_id: { type: string }
...
    Notification:
      type: object
      required: [id, organization_id, recipient_user_id, event_type, title, body, metadata, is_read, created_at]
      properties:
        id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        recipient_user_id: { type: string, format: uuid }
        event_type: { type: string, enum: [attempt.graded, assessment.published, resource.published] }
        title: { type: string }
        body: { type: string }
        metadata:
          type: object
          additionalProperties: { type: [string, number, boolean, "null"] }
        is_read: { type: boolean }
        read_at: { type: string, format: date-time, nullable: true }
        created_at: { type: string, format: date-time }
...
    NotificationUnreadCount:
      type: object
      required: [data]
      properties:
        data:
          type: object
          required: [count]
          properties:
            count: { type: integer, minimum: 0 }
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Forbidden:
      description: Insufficient permissions or invalid CSRF token
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
...